+++ database-backups/endereco.sql
@@ -89,7 +89,7 @@ CREATE TABLE `auth_permission` (
   PRIMARY KEY (`id`),
   UNIQUE KEY `auth_permission_content_type_id_codename_01ab375a_uniq` (`content_type_id`,`codename`),
   CONSTRAINT `auth_permission_content_type_id_2f476e4b_fk_django_co` FOREIGN KEY (`content_type_id`) REFERENCES `django_content_type` (`id`)
-) ENGINE=InnoDB AUTO_INCREMENT=25 DEFAULT CHARSET=latin1;
+) ENGINE=InnoDB AUTO_INCREMENT=34 DEFAULT CHARSET=latin1;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
 --
@@ -98,7 +98,7 @@ CREATE TABLE `auth_permission` (
 
 LOCK TABLES `auth_permission` WRITE;
 /*!40000 ALTER TABLE `auth_permission` DISABLE KEYS */;
-INSERT INTO `auth_permission` VALUES (1,'Can add log entry',1,'add_logentry'),(2,'Can change log entry',1,'change_logentry'),(3,'Can delete log entry',1,'delete_logentry'),(4,'Can add permission',2,'add_permission'),(5,'Can change permission',2,'change_permission'),(6,'Can delete permission',2,'delete_permission'),(7,'Can add group',3,'add_group'),(8,'Can change group',3,'change_group'),(9,'Can delete group',3,'delete_group'),(10,'Can add content type',4,'add_contenttype'),(11,'Can change content type',4,'change_contenttype'),(12,'Can delete content type',4,'delete_contenttype'),(13,'Can add session',5,'add_session'),(14,'Can change session',5,'change_session'),(15,'Can delete session',5,'delete_session'),(16,'Can add user',6,'add_user'),(17,'Can change user',6,'change_user'),(18,'Can delete user',6,'delete_user'),(19,'Can add entidade',7,'add_entity'),(20,'Can change entidade',7,'change_entity'),(21,'Can delete entidade',7,'delete_entity'),(22,'Can add Endereço',8,'add_endereco'),(23,'Can change Endereço',8,'change_endereco'),(24,'Can delete Endereço',8,'delete_endereco');
+INSERT INTO `auth_permission` VALUES (1,'Can add log entry',1,'add_logentry'),(2,'Can change log entry',1,'change_logentry'),(3,'Can delete log entry',1,'delete_logentry'),(4,'Can add permission',2,'add_permission'),(5,'Can change permission',2,'change_permission'),(6,'Can delete permission',2,'delete_permission'),(7,'Can add group',3,'add_group'),(8,'Can change group',3,'change_group'),(9,'Can delete group',3,'delete_group'),(10,'Can add content type',4,'add_contenttype'),(11,'Can change content type',4,'change_contenttype'),(12,'Can delete content type',4,'delete_contenttype'),(13,'Can add session',5,'add_session'),(14,'Can change session',5,'change_session'),(15,'Can delete session',5,'delete_session'),(16,'Can add user',6,'add_user'),(17,'Can change user',6,'change_user'),(18,'Can delete user',6,'delete_user'),(19,'Can add entidade',7,'add_entity'),(20,'Can change entidade',7,'change_entity'),(21,'Can delete entidade',7,'delete_entity'),(22,'Can add Endereço',8,'add_endereco'),(23,'Can change Endereço',8,'change_endereco'),(24,'Can delete Endereço',8,'delete_endereco'),(25,'Can add Pessoa',9,'add_pessoa'),(26,'Can change Pessoa',9,'change_pessoa'),(27,'Can delete Pessoa',9,'delete_pessoa'),(28,'Can add Fornecedor',10,'add_fornecedor'),(29,'Can change Fornecedor',10,'change_fornecedor'),(30,'Can delete Fornecedor',10,'delete_fornecedor'),(31,'Can add Cliente',11,'add_cliente'),(32,'Can change Cliente',11,'change_cliente'),(33,'Can delete Cliente',11,'delete_cliente');
 /*!40000 ALTER TABLE `auth_permission` ENABLE KEYS */;
 UNLOCK TABLES;
 
@@ -146,7 +146,7 @@ CREATE TABLE `django_content_type` (
   `model` varchar(100) NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE KEY `django_content_type_app_label_model_76bd3d3b_uniq` (`app_label`,`model`)
-) ENGINE=InnoDB AUTO_INCREMENT=9 DEFAULT CHARSET=latin1;
+) ENGINE=InnoDB AUTO_INCREMENT=12 DEFAULT CHARSET=latin1;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
 --
@@ -155,7 +155,7 @@ CREATE TABLE `django_content_type` (
 
 LOCK TABLES `django_content_type` WRITE;
 /*!40000 ALTER TABLE `django_content_type` DISABLE KEYS */;
-INSERT INTO `django_content_type` VALUES (1,'admin','logentry'),(3,'auth','group'),(2,'auth','permission'),(4,'contenttypes','contenttype'),(7,'fe_core','entity'),(6,'fe_core','user'),(8,'fe_endereco','endereco'),(5,'sessions','session');
+INSERT INTO `django_content_type` VALUES (1,'admin','logentry'),(3,'auth','group'),(2,'auth','permission'),(4,'contenttypes','contenttype'),(7,'fe_core','entity'),(6,'fe_core','user'),(8,'fe_endereco','endereco'),(11,'fe_pessoa','cliente'),(10,'fe_pessoa','fornecedor'),(9,'fe_pessoa','pessoa'),(5,'sessions','session');
 /*!40000 ALTER TABLE `django_content_type` ENABLE KEYS */;
 UNLOCK TABLES;
 
@@ -171,7 +171,7 @@ CREATE TABLE `django_migrations` (
   `name` varchar(255) NOT NULL,
   `applied` datetime(6) NOT NULL,
   PRIMARY KEY (`id`)
-) ENGINE=InnoDB AUTO_INCREMENT=17 DEFAULT CHARSET=latin1;
+) ENGINE=InnoDB AUTO_INCREMENT=20 DEFAULT CHARSET=latin1;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
 --
@@ -180,7 +180,7 @@ CREATE TABLE `django_migrations` (
 
 LOCK TABLES `django_migrations` WRITE;
 /*!40000 ALTER TABLE `django_migrations` DISABLE KEYS */;
-INSERT INTO `django_migrations` VALUES (1,'contenttypes','0001_initial','2018-06-14 21:56:24.715596'),(2,'contenttypes','0002_remove_content_type_name','2018-06-14 21:56:24.795667'),(3,'auth','0001_initial','2018-06-14 21:56:25.067123'),(4,'auth','0002_alter_permission_name_max_length','2018-06-14 21:56:25.085010'),(5,'auth','0003_alter_user_email_max_length','2018-06-14 21:56:25.099346'),(6,'auth','0004_alter_user_username_opts','2018-06-14 21:56:25.114691'),(7,'auth','0005_alter_user_last_login_null','2018-06-14 21:56:25.130550'),(8,'auth','0006_require_contenttypes_0002','2018-06-14 21:56:25.136642'),(9,'auth','0007_alter_validators_add_error_messages','2018-06-14 21:56:25.150299'),(10,'auth','0008_alter_user_username_max_length','2018-06-14 21:56:25.164346'),(11,'fe_core','0001_initial','2018-06-14 21:56:25.611756'),(12,'admin','0001_initial','2018-06-14 21:56:25.763306'),(13,'admin','0002_logentry_remove_auto_add','2018-06-14 21:56:25.785589'),(14,'auth','0009_alter_user_last_name_max_length','2018-06-14 21:56:25.806431'),(15,'fe_endereco','0001_initial','2018-06-14 21:56:25.952381'),(16,'sessions','0001_initial','2018-06-14 21:56:26.002682');
+INSERT INTO `django_migrations` VALUES (1,'contenttypes','0001_initial','2018-06-14 21:56:24.715596'),(2,'contenttypes','0002_remove_content_type_name','2018-06-14 21:56:24.795667'),(3,'auth','0001_initial','2018-06-14 21:56:25.067123'),(4,'auth','0002_alter_permission_name_max_length','2018-06-14 21:56:25.085010'),(5,'auth','0003_alter_user_email_max_length','2018-06-14 21:56:25.099346'),(6,'auth','0004_alter_user_username_opts','2018-06-14 21:56:25.114691'),(7,'auth','0005_alter_user_last_login_null','2018-06-14 21:56:25.130550'),(8,'auth','0006_require_contenttypes_0002','2018-06-14 21:56:25.136642'),(9,'auth','0007_alter_validators_add_error_messages','2018-06-14 21:56:25.150299'),(10,'auth','0008_alter_user_username_max_length','2018-06-14 21:56:25.164346'),(11,'fe_core','0001_initial','2018-06-14 21:56:25.611756'),(12,'admin','0001_initial','2018-06-14 21:56:25.763306'),(13,'admin','0002_logentry_remove_auto_add','2018-06-14 21:56:25.785589'),(14,'auth','0009_alter_user_last_name_max_length','2018-06-14 21:56:25.806431'),(15,'fe_endereco','0001_initial','2018-06-14 21:56:25.952381'),(16,'sessions','0001_initial','2018-06-14 21:56:26.002682'),(17,'fe_pessoa','0001_initial','2018-07-15 05:46:12.300349'),(18,'fe_pessoa','0002_auto_20170616_1501','2018-07-15 05:46:12.503053'),(19,'fe_pessoa','0003_auto_20170627_1615','2018-07-15 05:46:12.772418');
 /*!40000 ALTER TABLE `django_migrations` ENABLE KEYS */;
 UNLOCK TABLES;
 
@@ -354,6 +354,93 @@ LOCK TABLES `fe_endereco_endereco` WRITE;
 /*!40000 ALTER TABLE `fe_endereco_endereco` DISABLE KEYS */;
 /*!40000 ALTER TABLE `fe_endereco_endereco` ENABLE KEYS */;
 UNLOCK TABLES;
+
+--
+-- Table structure for table `fe_pessoa_cliente`
+--
+
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `fe_pessoa_cliente` (
+  `pessoa_ptr_id` char(32) NOT NULL,
+  PRIMARY KEY (`pessoa_ptr_id`),
+  CONSTRAINT `fe_pessoa_cliente_pessoa_ptr_id_86ce18a1_fk_fe_pessoa` FOREIGN KEY (`pessoa_ptr_id`) REFERENCES `fe_pessoa_pessoa` (`uuid`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Dumping data for table `fe_pessoa_cliente`
+--
+
+LOCK TABLES `fe_pessoa_cliente` WRITE;
+/*!40000 ALTER TABLE `fe_pessoa_cliente` DISABLE KEYS */;
+/*!40000 ALTER TABLE `fe_pessoa_cliente` ENABLE KEYS */;
+UNLOCK TABLES;
+
+--
+-- Table structure for table `fe_pessoa_fornecedor`
+--
+
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `fe_pessoa_fornecedor` (
+  `pessoa_ptr_id` char(32) NOT NULL,
+  PRIMARY KEY (`pessoa_ptr_id`),
+  CONSTRAINT `fe_pessoa_fornecedor_pessoa_ptr_id_08d1fe7d_fk_fe_pessoa` FOREIGN KEY (`pessoa_ptr_id`) REFERENCES `fe_pessoa_pessoa` (`uuid`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Dumping data for table `fe_pessoa_fornecedor`
+--
+
+LOCK TABLES `fe_pessoa_fornecedor` WRITE;
+/*!40000 ALTER TABLE `fe_pessoa_fornecedor` DISABLE KEYS */;
+/*!40000 ALTER TABLE `fe_pessoa_fornecedor` ENABLE KEYS */;
+UNLOCK TABLES;
+
+--
+-- Table structure for table `fe_pessoa_pessoa`
+--
+
+/*!40101 SET @saved_cs_client     = @@character_set_client */;
+/*!40101 SET character_set_client = utf8 */;
+CREATE TABLE `fe_pessoa_pessoa` (
+  `uuid` char(32) NOT NULL,
+  `created_at` datetime(6) NOT NULL,
+  `updated_at` datetime(6) NOT NULL,
+  `transiente` tinyint(1) NOT NULL,
+  `entidade_id` char(32) DEFAULT NULL,
+  `nome` varchar(128) DEFAULT NULL,
+  `cpf` varchar(14) DEFAULT NULL,
+  `nome_fantasia` varchar(128) DEFAULT NULL,
+  `contato` varchar(128) DEFAULT NULL,
+  `data_nascimento` date DEFAULT NULL,
+  `inscricao_estadual` varchar(20) DEFAULT NULL,
+  `inscricao_municipal` varchar(20) DEFAULT NULL,
+  `endereco` char(32) DEFAULT NULL,
+  `telefone_celular` varchar(32) DEFAULT NULL,
+  `telefone_comercial` varchar(32) DEFAULT NULL,
+  `telefone_residencial` varchar(32) DEFAULT NULL,
+  `telefone_recado` varchar(32) DEFAULT NULL,
+  `site` varchar(200) DEFAULT NULL,
+  `email` varchar(254) DEFAULT NULL,
+  `observacao` longtext,
+  `ativo` tinyint(1) NOT NULL,
+  PRIMARY KEY (`uuid`),
+  KEY `fe_pessoa_pessoa_entidade_id_b9d43c1f` (`entidade_id`),
+  CONSTRAINT `fe_pessoa_pessoa_entidade_id_b9d43c1f_fk_fe_core_entity_uuid` FOREIGN KEY (`entidade_id`) REFERENCES `fe_core_entity` (`uuid`)
+) ENGINE=InnoDB DEFAULT CHARSET=latin1;
+/*!40101 SET character_set_client = @saved_cs_client */;
+
+--
+-- Dumping data for table `fe_pessoa_pessoa`
+--
+
+LOCK TABLES `fe_pessoa_pessoa` WRITE;
+/*!40000 ALTER TABLE `fe_pessoa_pessoa` DISABLE KEYS */;
+/*!40000 ALTER TABLE `fe_pessoa_pessoa` ENABLE KEYS */;
+UNLOCK TABLES;
 /*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
 
 /*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
@@ -364,4 +451,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2018-07-03 14:14:42
+-- Dump completed on 2018-07-15  5:47:19